+++ engine/cld/legacy/cli/mcmsv2/err_decode_helpers.go
@@ -1,4 +1,4 @@
-package common
+package mcmsv2
 
 import (
 	"bytes"
