+++ engine/cld/legacy/cli/mcmsv2/helpers.go
@@ -1,4 +1,4 @@
-package common
+package mcmsv2
 
 import (
 	"errors"
