+++ engine/cld/legacy/cli/mcmsv2/mcms_v2.go
@@ -1,4 +1,4 @@
-package common
+package mcmsv2
 
 import (
 	"context"
