+++ engine/cld/legacy/cli/mcmsv2/mcms_v2_test.go
@@ -1,4 +1,4 @@
-package common
+package mcmsv2
 
 import (
 	"context"
